Add solver
OK this is a big push so I should probably do some explaining on the algorithm huh.
First, there's a weight function (dictionary) that maps every English letter to a float.
This float (taken from Wikipedia) indicates the relative frequency of each letter in dictionaries, as a percentage.
This mapping is used in the weigh_word() function, which uses that in combination with
the frequency of the letter in the word itself to come up with a numerical weight of the word.
The reason frequency within the word is so crucial is because repeated letters make for bad guesses,
and making duplicates weigh less alleviates this problem (also before, it claimed the best word was 'reese').
The solve_step() function takes a word and the given hints, and computes the next best guess to make.
These calculations are done by removing potential entries that no longer align with given hint criteria.
The final solve_puzzle() function combines these two methods, prompting user input.
The user is told to enter the next best guess, and upon doing so, they will be prompted whether they won the game.
If not, they will be asked for five letters relating to the given hints.
These are run through the solve_step() once again, until another possible solution is given.
I still need to run some more testing, but it should solve most puzzles in 3 or 4 guesses.
I'll report back with test data in the next commit. Over and out!