Add tester for solutions; update menu
As promised, here are the test details:
Wordles guess in 1 try:   1 (arise)
Wordles guess in 2 try:  43
Wordles guess in 3 try: 393
Wordles guess in 4 try: 784
Wordles guess in 5 try: 588
Wordles guess in 6 try: 292
Wordles left unsolved:  214
So that's about a 10% discrepancy, but it still seems logically sound!
I also tried messing around with the weighting of letters, but it only
ever made the results worse.